feat: Implement ComparativeAnalysis component in NewApprovalForm for enhanced comparison rate handling; refactor related logic and remove redundant state management.
